requirements.txt
opencv-python>=4.9.0.80
keyboard>=0.13.5
mouse>=0.7.1
numpy>=1.26.4
psutil>=5.9.8
pillow>=10.3.0
pyinstaller>=6.6.0
